+++ jib-core/src/integration-test/java/com/google/cloud/tools/jib/registry/AuthenticationMethodRetrieverIntegrationTest.java
@@ -17,7 +17,9 @@
 package com.google.cloud.tools.jib.registry;
 
 import com.google.cloud.tools.jib.http.Authorization;
+import com.google.cloud.tools.jib.http.Connection;
 import java.io.IOException;
+import org.junit.Assert;
 import org.junit.Test;
 
 /** Integration tests for {@link AuthenticationMethodRetriever}. */
@@ -27,14 +29,17 @@ public class AuthenticationMethodRetrieverIntegrationTest {
   public void testGetRegistryAuthenticator()
       throws RegistryAuthenticationFailedException, IOException, RegistryException {
     RegistryClient registryClient =
-        RegistryClient.factory("registry.hub.docker.com", "library/busybox", null)
-            .newWithAuthorization(null);
+        RegistryClient.factory(Connection::new,
+            "registry.hub.docker.com", "library/busybox").newRegistryClient();
     RegistryAuthenticator registryAuthenticator = registryClient.getRegistryAuthenticator();
+    Assert.assertNotNull(registryAuthenticator);
     Authorization authorization = registryAuthenticator.authenticatePull();
 
     RegistryClient authorizedRegistryClient =
-        RegistryClient.factory("registry.hub.docker.com", "library/busybox", null)
-            .newWithAuthorization(authorization);
+        RegistryClient.factory(Connection::new,
+            "registry.hub.docker.com", "library/busybox")
+            .setAuthorization(authorization)
+            .newRegistryClient();
     authorizedRegistryClient.pullManifest("latest");
   }
 }